backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
chromadb
PyPDF2
python-docx
google-generativeai
python-dotenv
pyjwt
sentence-transformers
torch